Move TestUpdateMethods to handler package for clarity
