+++ gengokit/generator/gen_test.go
@@ -10,10 +10,8 @@ import (
 	"testing"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/pkg/errors"
 
 	"github.com/TuneLab/go-truss/gengokit"
-	"github.com/TuneLab/go-truss/gengokit/handler"
 	templateFileAssets "github.com/TuneLab/go-truss/gengokit/template"
 	"github.com/TuneLab/go-truss/svcdef"
 
@@ -270,170 +268,6 @@ func TestAllTemplates(t *testing.T) {
 				templFP, err, addRPCCode)
 		}
 	}
-}
-
-func TestUpdateMethods(t *testing.T) {
-	const def = `
-		syntax = "proto3";
-
-		// General package
-		package general;
-
-		import "google.golang.org/genproto/googleapis/api/serviceconfig/annotations.proto";
-
-		// RequestMessage is so foo
-		message RequestMessage {
-			string input = 1;
-		}
-
-		// ResponseMessage is so bar
-		message ResponseMessage {
-			string output = 1;
-		}
-
-		// RequestMessage is so foo
-		message DifferentRequest {
-			float green = 1;
-		}
-
-		message DifferentResponse {
-			int64 blue = 1;
-		}
-
-		// ProtoService is a service
-		service ProtoService {
-			// ProtoMethod is simple. Like a gopher.
-			rpc ProtoMethod (RequestMessage) returns (ResponseMessage) {
-				// No {} in path and no body, everything is in the query
-				option (google.api.http) = {
-					get: "/route"
-				};
-			}
-			// ProtoMethodAgain is simple. Like a gopher again.
-			rpc ProtoMethodAgain (RequestMessage) returns (ResponseMessage) {
-				// No {} in path and no body, everything is in the query
-				option (google.api.http) = {
-					get: "/route2"
-				};
-			}
-			// ProtoMethodAgainAgain is simple. Like a gopher again again.
-			rpc ProtoMethodAgainAgain (RequestMessage) returns (ResponseMessage) {
-				// No {} in path and no body, everything is in the query
-				option (google.api.http) = {
-					get: "/route2"
-				};
-			}
-		}
-	`
-
-	sd, err := svcdef.NewFromString(def, gopath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	svc := sd.Service
-	allMethods := svc.Methods
-
-	conf := gengokit.Config{
-		GoPackage: "github.com/TuneLab/go-truss/gengokit",
-		PBPackage: "github.com/TuneLab/go-truss/gengokit/general-service",
-	}
-
-	te, err := gengokit.NewData(sd, conf)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	templPath := handler.ServerHandlerPath
-
-	svc.Methods = []*svcdef.ServiceMethod{allMethods[0]}
-
-	firstCode, err := renderService(svc, "", te, templPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	secondCode, err := renderService(svc, firstCode, te, templPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(firstCode) != len(secondCode) {
-		t.Fatal("Generated code differs after regenerated with same definition\n" +
-			templPath + "\n" +
-			diff(firstCode, secondCode))
-	}
-
-	svc.Methods = append(svc.Methods, allMethods[1])
-
-	thirdCode, err := renderService(svc, secondCode, te, templPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(thirdCode) <= len(secondCode) {
-		t.Fatal("Generated code not longer after regenerated with additional service method\n" +
-			templPath + "\n" +
-			diff(secondCode, thirdCode))
-	}
-
-	// remove the first one rpc
-	svc.Methods = svc.Methods[1:]
-
-	fourthCode, err := renderService(svc, thirdCode, te, templPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(fourthCode) >= len(thirdCode) {
-		t.Fatal("Generated code not shorter after regenerated with fewer service method\n" +
-			templPath + "\n" +
-			diff(thirdCode, fourthCode))
-	}
-
-	svc.Methods = allMethods
-
-	fifthCode, err := renderService(svc, fourthCode, te, templPath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(fifthCode) <= len(fourthCode) {
-		t.Fatal("Generated code not longer after regenerated with additional service method\n" +
-			templPath + "\n" +
-			diff(fourthCode, fifthCode))
-	}
-}
-
-func renderService(svc *svcdef.Service, prevFileAsString string, te *gengokit.Data, templPath string) (string, error) {
-	var prevFile io.Reader
-	if prevFileAsString != "" {
-		prevFile = strings.NewReader(prevFileAsString)
-	}
-
-	h, err := handler.New(svc, prevFile, te.PackageName)
-	if err != nil {
-		return "", err
-	}
-
-	next, err := h.Render(templPath, te)
-	if err != nil {
-		return "", err
-	}
-
-	nextBytes, err := ioutil.ReadAll(next)
-	if err != nil {
-		return "", err
-	}
-
-	nextCode, err := testFormat(string(nextBytes))
-	if err != nil {
-		return "", errors.Wrap(err, "cannot format")
-	}
-
-	nextCode = strings.TrimSpace(nextCode)
-
-	return nextCode, nil
 }
 
 func diff(a, b string) string {
