+++ gengokit/gentesthelper/helper.go
@@ -92,3 +92,15 @@ func DiffGoCode(inA, inB string) (outA, outB, diff string) {
 	)
 	return
 }
+
+// testFormat takes a string representing golang code and attempts to return a
+// formated copy of that code.
+func TestFormat(code string) (string, error) {
+	formatted, err := format.Source([]byte(code))
+
+	if err != nil {
+		return code, err
+	}
+
+	return string(formatted), nil
+}
